requirements.txt
APScheduler==3.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
dataclasses-json==0.5.7
deep-translator==1.8.3
dnspython==2.2.1
Flask==2.1.2
futures==2.2.0
googletrans==3.0.0
goslate==1.5.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.9.1
pymongo==4.2.0
pyparsing==3.0.9
pyshorteners==1.0.1
pyTelegramBotAPI==4.6.0
pytest==7.1.2
python-dotenv==0.20.0
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stripe==3.5.0
Telethon==1.24.0
tomli==2.0.1
tornado==6.1
typing-inspect==0.7.1
typing_extensions==4.2.0
tzdata==2021.5
tzlocal==4.1
ujson==5.3.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
